backend/requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.32.1
motor==3.6.0
pymongo==4.9.2
pydantic[email]==2.10.3
pydantic-settings==2.6.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
python-multipart==0.0.18
groq==0.13.0
python-dotenv==1.0.1
